back-end/requirements.txt
Django==5.0.3
djangorestframework==3.15.0
djangorestframework_simplejwt
pillow==10.2.0